Add SUBSYSTEM check to udev rules
Removing this causes issues with ttyUSB devices as there are two of them on the
device tree: one at the top with tty subsystem that we want, and one below it
with usb-serial subsystem. Previous version started the service for both of
these devices, with usb-serial subsystem device failing immediately. This commit
fixes this issue.